+++ src/main/java/basemod/abstracts/CustomUnlock.java
@@ -1,15 +1,27 @@
 package basemod.abstracts;
 
 import com.megacrit.cardcrawl.helpers.CardLibrary;
+import com.megacrit.cardcrawl.helpers.RelicLibrary;
 import com.megacrit.cardcrawl.unlock.AbstractUnlock;
 
 public class CustomUnlock extends AbstractUnlock {
 
 	public CustomUnlock(String cardID) {
-		this.type = AbstractUnlock.UnlockType.CARD;
-		this.card = CardLibrary.getCard(cardID);
-		this.key = this.card.cardID;
-		this.title = this.card.name;
+		this(UnlockType.CARD, cardID);
+	}
+
+	public CustomUnlock(UnlockType type, String id) {
+		this.type = type;
+		switch (this.type) {
+			case CARD:
+				card = CardLibrary.getCard(id);
+				key = card.cardID;
+				title = card.name;
+			case RELIC:
+				relic = RelicLibrary.getRelic(id);
+				key = relic.relicId;
+				title = relic.name;
+		}
 	}
 	
 }