+++ CHANGELOG.md
@@ -413,3 +413,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### dev ####
 * Fix Start Game hook triggering when looping in Endless mode (kiooeht)
 * Fix Ascension 20 double boss when custom bosses have been added (kiooeht)
+* Allow custom characters to add relic unlocks (kiooeht)